modelcheck and ifc  splitted
