+++ som_gui/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "2.4.4"
+__version__ = "2.6.0"
